Updating to latest FlutterFlow output.

+++ lib/flutter_flow/custom_functions.dart
@@ -17,7 +17,7 @@ double soma(
   double num1,
   double num2,
 ) {
-  return num1 * num2;
+  return num1 * num2.toDouble();
 }
 
 double subtracao(double valor1) {
